renamed the source file containing the model to model
