+++ train_model.py
@@ -6,7 +6,7 @@ import math
 from torch.utils.tensorboard import SummaryWriter
 
 from ml_model.dataset import CustomDataset
-from ml_model.neural_network import NeuralNetwork
+from ml_model.model import NeuralNetwork
 
 TRAINED_MODEL_DIR = "trained_models/"
